renaming things to match standard
