+++ src/app/(main)/page.tsx
@@ -1,8 +1,8 @@
 import Image from 'next/image';
 
 import { AppContentWrapper } from '@/components/common/Global.style';
-import Hero from '@/components/Home/Hero/Hero';
-import Stats from '@/components/Home/Stats/Stats';
+import Hero from '@/components/home/hero/Hero';
+import Stats from '@/components/home/stats/Stats';
 
 export default function Home() {
   return (
@@ -12,7 +12,7 @@ export default function Home() {
       <AppContentWrapper>
         <Stats />
 
-        <Image src={'/Decorative/crosses.svg'} width={14} height={80} alt="crosses" style={{ float: 'right' }} />
+        <Image src={'/decorative/crosses.svg'} width={14} height={80} alt="crosses" style={{ float: 'right' }} />
 
       </AppContentWrapper>
     </>
